change FPS values to double precision

+++ lib608/608.c
@@ -15,8 +15,8 @@ const timecode default_timecode = {0, 0, 0, 0, false};
 const VersionInfo library_version = {0, 5, 0, 1, VERSION};
 
 // no idea why the 1/2 frame is needed but idk
-s64 tc2int(timecode pts, f32 fps) {
-	f32 ret;
+s64 tc2int(timecode pts, f64 fps) {
+	f64 ret;
 	if (pts.drop) {
 		ret = (((60*60*fps)-108)*pts.hours)+(((60*10*fps)-18)*(pts.minutes/10.0f))+(((60*fps)-2)*(fmodf(pts.minutes, 10)))+(fps*pts.seconds)+pts.frames+0.5f;
 	}
@@ -26,11 +26,11 @@ s64 tc2int(timecode pts, f32 fps) {
 	log_write(LOG_TRACE, use_colors, "tc2int: %02d:%02hhu:%02hhu%c%02hhu -> 0x%08x\n", pts.hours, pts.minutes, pts.seconds, pts.drop ? ';' : ':', pts.frames, (s64) ret);
 	return (s64) ret;
 }
-timecode int2tc(s64 pts, f32 fps, bool8 drop) {
+timecode int2tc(s64 pts, f64 fps, bool8 drop) {
 	timecode ret = {0};
-	f32 _pts = (f32) pts; // for logging pruposes
+	f64 _pts = (f64) pts; // for logging pruposes
 	// McPoodle's tools fail on negative pts; however, NW4R format allows for it so properly handle it
-	f32 hours, minutes, seconds, frames, dm; // To allow operations with the fps value
+	f64 hours, minutes, seconds, frames, dm; // To allow operations with the fps value
 	if (drop) {
 		hours = (_pts+108)/(60*60*fps);
 		if (_pts < 0) {
